+++ be/src/exec/olap_scan_prepare.cpp
@@ -36,7 +36,7 @@
 #include "storage/runtime_filter_predicate.h"
 #include "storage/runtime_range_pruner.h"
 #include "storage/runtime_range_pruner.hpp"
-#include "types/date_value.hpp"
+#include "types/date_value.h"
 #include "types/logical_type.h"
 #include "types/logical_type_infra.h"
