+++ be/src/formats/csv/converter.h
@@ -26,7 +26,7 @@
 #include "base/string/slice.h"
 #include "common/status.h"
 #include "io/formatted_output_stream.h"
-#include "types/date_value.hpp"
+#include "types/date_value.h"
 #include "types/decimalv2_value.h"
 #include "types/time_types.h"
 #include "types/timestamp_value.h"
