+++ be/src/storage/row_store_encoder_simple.cpp
@@ -26,7 +26,7 @@
 #include "storage/primary_key_encoder.h"
 #include "storage/row_store_encoder_util.h"
 #include "storage/tablet_schema.h"
-#include "types/date_value.hpp"
+#include "types/date_value.h"
 
 namespace starrocks {
