+++ be/src/storage/rowset/bitshuffle_page.h
@@ -56,7 +56,7 @@
 #include "storage/rowset/page_builder.h"
 #include "storage/rowset/page_decoder.h"
 #include "storage/types.h"
-#include "types/date_value.hpp"
+#include "types/date_value.h"
 #include "types/type_traits.h"
 
 namespace starrocks {
