+++ be/src/types/date_value.h
@@ -136,6 +136,18 @@ inline int32_t DateValue::to_days_since_unix_epoch() const {
     return _julian - date::UNIX_EPOCH_JULIAN;
 }
 
+inline void DateValue::to_date(int* year, int* month, int* day) const {
+    date::to_date_with_cache(_julian, year, month, day);
+}
+
+inline std::chrono::sys_days DateValue::to_sys_days() const {
+    int year, month, day;
+    to_date(&year, &month, &day);
+    return std::chrono::sys_days{std::chrono::year_month_day{std::chrono::year{year},
+                                                             std::chrono::month{static_cast<unsigned>(month)},
+                                                             std::chrono::day{static_cast<unsigned>(day)}}};
+}
+
 template <TimeUnit UNIT>
 DateValue DateValue::add(int count) const {
     return DateValue{date::add<UNIT>(_julian, count)};
