+++ be/test/types/date_value_test.cpp
@@ -20,7 +20,6 @@
 #include <type_traits>
 
 #include "types/date_value.h"
-#include "types/date_value.hpp"
 #include "types/time_types.h"
 #include "types/timestamp_value.h"
