+++ be/test/types/timestamp_value_test.cpp
@@ -17,7 +17,7 @@
 
 #define private public
 
-#include "types/date_value.hpp"
+#include "types/date_value.h"
 #include "types/time_types.h"
 #include "types/timestamp_value.h"
